Squashed commit of the following:

commit 2dd056ac9f6e65d99293d914ebc505636d8ba03a
Author: Sunirmal Mukherjee <sunimukh@cisco.com>
Date:   Tue Sep 1 03:10:30 2020 -0400

    build: Updates for libdaq changes to interface group field width and naming

commit ede03a067cf4c5fa98321453d97a1cda544c6b1d
Author: Sunirmal Mukherjee <sunimukh@cisco.com>
Date:   Tue Sep 1 03:09:18 2020 -0400

    helpers: Rework DiscoveryFilter to monitor IP lists based on interface rather than group

commit dc80cdbeb1b12913b5f3968a49e8cc5db5da6d2e
Author: Sunirmal Mukherjee <sunimukh@cisco.com>
Date:   Tue Sep 1 02:56:54 2020 -0400

    reputation: Change from group-based to interface-based IP lists

+++ src/stream/tcp/tcp_stream_session.h
@@ -119,9 +119,9 @@ public:
     uint32_t pkt_action_mask = ACTION_NOTHING;
     uint8_t ecn = 0;
     int32_t ingress_index = 0;
-    int32_t ingress_group = 0;
+    int16_t ingress_group = 0;
     int32_t egress_index = 0;
-    int32_t egress_group = 0;
+    int16_t egress_group = 0;
     uint32_t daq_flags = 0;
     uint16_t address_space_id = 0;
     bool generate_3whs_alert = true;
